Add ITitleScreenMenu and Scoped Service.
